Correct address validation
Correct validation such that if the geocoded value is not valid, unset
the value of the field. If the field is valid, overwrite the value with
the returned formatted value, since that's what we're actually storing.
Finally, if the field blurs, just revert it back to the most recent
valid address.